+++ packages/amplify-e2e-core/src/utils/retrier.ts
@@ -11,36 +11,54 @@ const defaultSettings: RetrySettings = {
 /**
  * Retries the function func until the predicate pred returns true, or until one of the retry limits is met.
  * @param func The function to retry
- * @param pred The predicate that determines successful output of func
+ * @param successPredicate The predicate that determines successful output of func
  * @param settings Retry limits (defaults to defaultSettings above)
+ * @param failurePredicate An optional predicate that determines that the retry operation has failed and should not be retried anymore
  */
-export const retry = async <T>(func: () => Promise<T>, pred: (res?: T) => boolean, settings?: Partial<RetrySettings>) => {
-  const { times, delayMS, timeoutMS, stopOnError } = _.merge({}, defaultSettings, settings);
+export const retry = async <T>(
+  func: () => Promise<T>,
+  successPredicate: (res?: T) => boolean,
+  settings?: Partial<RetrySettings>,
+  failurePredicate?: (res?: T) => boolean,
+): Promise<T> => {
+  const {
+    times, delayMS, timeoutMS, stopOnError,
+  } = _.merge({}, defaultSettings, settings);
 
   let count = 0;
-  let result: T = undefined;
+  let result: T;
   let terminate = false;
   const startTime = Date.now();
 
   do {
     try {
       result = await func();
-      if (pred(result)) {
+      if (successPredicate(result)) {
         return result;
-      } else {
-        console.warn(`Retryable function execution did not match predicate. Result was [${JSON.stringify(result)}]. Retrying...`);
       }
+      if (typeof failurePredicate === 'function' && failurePredicate(result)) {
+        throw new Error('Retry-able function execution result matched failure predicate. Stopping retries.');
+      }
+      console.warn(`Retry-able function execution did not match success predicate. Result was [${JSON.stringify(result)}]. Retrying...`);
     } catch (err) {
-      console.warn(`Retryable function execution failed with [${err}]. Retrying...`);
+      console.warn(`Retry-able function execution failed with [${err.message || err}]`);
+      if (stopOnError) {
+        console.log('Stopping retries on error.');
+      } else {
+        console.log('Retrying...');
+      }
       terminate = stopOnError;
     }
     count++;
     await sleep(delayMS);
   } while (!terminate && count <= times && Date.now() - startTime < timeoutMS);
 
-  throw new Error('Retryable function did not match predicate within the given retry constraints');
+  throw new Error('Retry-able function did not match predicate within the given retry constraints');
 };
 
+/**
+ * Configuration for retry limits
+ */
 export type RetrySettings = {
   times: number; // specifying 1 will execute func once and if not successful, retry one time
   delayMS: number; // delay between each attempt to execute func (there is no initial delay)