+++ packages/amplify-provider-awscloudformation/src/aws-utils/aws-amplify-backend.ts
@@ -1,13 +1,19 @@
 import AWS from 'aws-sdk';
+import { $TSContext } from 'amplify-cli-core';
 import aws from './aws';
 import { loadConfiguration } from '../configuration-manager';
-import { $TSContext } from 'amplify-cli-core';
 
+/**
+ * Amplify Backend Class
+ */
 export class AmplifyBackend {
   private static instance: AmplifyBackend;
   private readonly context: $TSContext;
   public amplifyBackend: AWS.AmplifyBackend;
 
+  /**
+   * Configures the instance for the Amplify Backend Client
+   */
   static async getInstance(context: $TSContext, options = {}): Promise<AmplifyBackend> {
     if (!AmplifyBackend.instance) {
       let cred = {};
@@ -23,6 +29,12 @@ export class AmplifyBackend {
 
   private constructor(context: $TSContext, creds, options = {}) {
     this.context = context;
-    this.amplifyBackend = new aws.AmplifyBackend({ ...creds, ...options });
+    const { AMPLIFY_BACKEND_ENDPOINT, AMPLIFY_BACKEND_REGION } = process.env;
+    this.amplifyBackend = new aws.AmplifyBackend({
+      ...creds,
+      ...options,
+      ...(AMPLIFY_BACKEND_ENDPOINT && { endpoint: AMPLIFY_BACKEND_ENDPOINT }),
+      ...(AMPLIFY_BACKEND_REGION && { region: AMPLIFY_BACKEND_REGION }),
+    });
   }
 }