+++ packages/amplify-provider-awscloudformation/src/utils/admin-helpers.ts
@@ -4,6 +4,7 @@ import _ from 'lodash';
 import fetch from 'node-fetch';
 import { adminLoginFlow } from '../admin-login';
 import { AdminAuthConfig, AwsSdkConfig, CognitoAccessToken, CognitoIdToken } from './auth-types';
+import proxyAgent from 'proxy-agent';
 
 export const adminVerifyUrl = (appId: string, envName: string, region: string): string => {
   const baseUrl = process.env.AMPLIFY_CLI_ADMINUI_BASE_URL ?? adminBackendMap[region]?.amplifyAdminUrl;
@@ -48,7 +49,9 @@ export async function getTempCredsWithAdminTokens(context: $TSContext, appId: st
 async function getAdminAppState(appId: string, region: string) {
   // environment variable AMPLIFY_CLI_APPSTATE_BASE_URL useful for development against beta/gamma appstate endpoints
   const appStateBaseUrl = process.env.AMPLIFY_CLI_APPSTATE_BASE_URL ?? adminBackendMap[region].appStateUrl;
-  const res = await fetch(`${appStateBaseUrl}/AppState/?appId=${appId}`);
+  const httpProxy = process.env.HTTP_PROXY || process.env.HTTPS_PROXY;
+  const fetchOptions = httpProxy ? { agent: proxyAgent(httpProxy) } : {};
+  const res = await fetch(`${appStateBaseUrl}/AppState/?appId=${appId}`, fetchOptions);
   return res.json();
 }