+++ packages/amplify-util-mock/src/api/api.ts
@@ -69,7 +69,10 @@ export class APITest {
       context.print.info(`AppSync Mock endpoint is running at ${this.appSyncSimulator.url}`);
       await this.startDDBListeners(context, appSyncConfig, false);
     } catch (e) {
-      context.print.error(`Failed to start API Mock endpoint ${e}`);
+      const errMessage = 'Failed to start API Mocking.';
+      context.print.error(errMessage + ' Running cleanup tasks.');
+      await this.stop(context);
+      throw new Error(`${errMessage} Reason: ${e?.message}`);
     }
   }