+++ packages/amplify-util-uibuilder/src/commands/utils/syncAmplifyUiBuilderComponents.ts
@@ -1,86 +1,29 @@
-import { getAmplifyUIBuilderService, Component, Theme } from './amplifyUiBuilderService';
 import { printer } from 'amplify-prompts';
-import { createUiBuilderComponent, createUiBuilderTheme } from './createUiBuilderComponent';
-import { getUiBuilderComponentsPath } from './getUiBuilderComponentsPath';
-import { extractArgs } from './extractArgs';
 import { $TSContext } from 'amplify-cli-core';
-export const getEnvName = (context: $TSContext, envName?: string) => {
-  const args = extractArgs(context);
-  return envName ? envName : args.environmentName ? args.environmentName : context.exeInfo.localEnvInfo.envName;
-};
+import {
+  StudioComponent, StudioTheme, GenericDataSchema,
+} from '@aws-amplify/codegen-ui';
+import { createUiBuilderComponent, createUiBuilderTheme } from './codegenResources';
+import { getUiBuilderComponentsPath } from './getUiBuilderComponentsPath';
 
-export const resolveAppId = async (context: $TSContext) => {
-  return await context.amplify.invokePluginMethod(context, 'awscloudformation', undefined, 'resolveAppId', [context]);
-};
-
-export const getAppId = async (context: $TSContext) => {
-  const appId = extractArgs(context).appId || (await resolveAppId(context));
-
-  if (!appId) {
-    throw new Error(
-      'Unable to sync Studio components since appId could not be determined. This can happen when you hit the soft limit of number of apps that you can have in Amplify console.',
-    );
-  }
-  return appId;
-};
-
-export async function listUiBuilderComponents(context: $TSContext, envName?: string): Promise<{ entities: Component[] }> {
-  const environmentName = getEnvName(context, envName);
-  const appId = await getAppId(context);
-
-  try {
-    const amplifyUIBuilder = await getAmplifyUIBuilderService(context, environmentName, appId);
-    let nextToken: string | undefined;
-    const uiBuilderComponents: Component[] = [];
-    do {
-      const response = await amplifyUIBuilder
-      .exportComponents({
-        appId,
-        environmentName,
-        nextToken
-      })
-      .promise();
-      uiBuilderComponents.push(...response.entities);
-      nextToken = response.nextToken;
-    } while (nextToken);
-    printer.debug(JSON.stringify(uiBuilderComponents, null, 2));
-    return { entities: uiBuilderComponents };
-  } catch (e) {
-    printer.debug(e);
-    throw e;
-  }
-}
-
-export async function listUiBuilderThemes(context: $TSContext, envName?: string): Promise<{ entities: Theme[] }> {
-  const environmentName = getEnvName(context, envName);
-  const appId = await getAppId(context);
-
-  try {
-    const amplifyUIBuilder = await getAmplifyUIBuilderService(context, environmentName, appId);
-    let nextToken: string | undefined;
-    const uiBuilderThemes: Theme[] = [];
-    do {
-      const response = await amplifyUIBuilder
-      .exportThemes({
-        appId,
-        environmentName,
-      })
-      .promise();
-      uiBuilderThemes.push(...response.entities);
-      nextToken = response.nextToken;
-    } while (nextToken);
-    printer.debug(JSON.stringify(uiBuilderThemes, null, 2));
-    return { entities: uiBuilderThemes };
-  } catch (e) {
-    printer.debug(e);
-    throw e;
-  }
-}
-
-export function generateUiBuilderComponents(context: $TSContext, componentSchemas: any[]) {
+/**
+ * Returns instances of StudioComponent from the component schemas
+ */
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+export const generateUiBuilderComponents = (context: $TSContext, componentSchemas: any[], dataSchema?: GenericDataSchema): ({
+  resultType: string;
+  component: StudioComponent;
+  schemaName?: undefined;
+  error?: undefined;
+} | {
+  resultType: string;
+  schemaName: string;
+  error: Error;
+  component?: undefined;
+})[] => {
   const componentResults = componentSchemas.map(schema => {
     try {
-      const component = createUiBuilderComponent(context, schema);
+      const component = createUiBuilderComponent(context, schema, dataSchema);
       return { resultType: 'SUCCESS', component };
     } catch (e) {
       printer.debug(`Failure caught processing ${schema.name}`);
@@ -95,9 +38,23 @@ export function generateUiBuilderComponents(context: $TSContext, componentSchema
     )}`,
   );
   return componentResults;
-}
+};
 
-export function generateUiBuilderThemes(context: $TSContext, themeSchemas: any[]) {
+/**
+ * Returns instances of StudioTheme from theme schemas
+ */
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+export const generateUiBuilderThemes = (context: $TSContext, themeSchemas: any[]): ({
+  resultType: string;
+  theme: StudioTheme;
+  schemaName?: undefined;
+  error?: undefined;
+} | {
+  resultType: string;
+  schemaName: string;
+  error: Error;
+  theme?: undefined;
+})[] => {
   const themeResults = themeSchemas.map(schema => {
     try {
       const theme = createUiBuilderTheme(context, schema);
@@ -113,4 +70,4 @@ export function generateUiBuilderThemes(context: $TSContext, themeSchemas: any[]
     `Generated ${themeResults.filter(result => result.resultType === 'SUCCESS').length} themes in ${getUiBuilderComponentsPath(context)}`,
   );
   return themeResults;
-}
+};